+++ Pixel/src/Platform/DirectX/Context/DirectXContext.cpp
@@ -589,7 +589,7 @@ namespace Pixel {
 		throw std::logic_error("The method or operation is not implemented.");
 	}
 
-	void DirectXContext::SetDynamicIB(size_t IndexCount, const uint64_t* IBData)
+	void DirectXContext::SetDynamicIB(size_t IndexCount, const uint16_t* IBData)
 	{
 		throw std::logic_error("The method or operation is not implemented.");
 	}
