+++ Pixel/src/Platform/DirectX/Context/GraphicsContext.h
@@ -54,7 +54,7 @@ namespace Pixel {
 		virtual void SetVertexBuffer(uint32_t Slot, const Ref<VBV> VBView) override;
 		virtual void SetVertexBuffers(uint32_t StartSlot, uint32_t Count, const std::vector<Ref<VBV>> VBViews) override;
 		virtual void SetDynamicVB(uint32_t Slot, size_t NumVertices, size_t VertexStride, const void* VBData) override;
-		virtual void SetDynamicIB(size_t IndexCount, const uint64_t* IBData) override;
+		virtual void SetDynamicIB(size_t IndexCount, const uint16_t* IBData) override;
 		virtual void SetDynamicSRV(uint32_t RootIndex, size_t BufferSize, const void* BufferData) override;
 
 		virtual void Draw(uint32_t VertexCount, uint32_t VertexStartOffset = 0) override;
